+++ test/src/test_circular_buffer.cpp
@@ -27,6 +27,7 @@ int main()
     CUN_UNITTEST_NAME(ut, "default parameter check");
     CUN_UNITTEST_EXEC(ut, CircularBuffer<uint8_t, 1> cb1);
     CUN_UNITTEST_EVAL(ut, cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 1);
     CUN_UNITTEST_EVAL(ut, cb1.max_size() == 1);
     CUN_UNITTEST_EVAL(ut, cb1.size() == 0);
     CUN_UNITTEST_NL(ut);
@@ -34,6 +35,7 @@ int main()
     CUN_UNITTEST_NAME(ut, "push data: part.1");
     CUN_UNITTEST_EVAL(ut, cb1.push(1));
     CUN_UNITTEST_EVAL(ut, !cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 0);
     CUN_UNITTEST_EVAL(ut, cb1.size() == 1);
     CUN_UNITTEST_NL(ut);
 
@@ -42,6 +44,7 @@ int main()
     CUN_UNITTEST_EVAL(ut, cb1.pop(val));
     CUN_UNITTEST_EVAL(ut, val == 1);
     CUN_UNITTEST_EVAL(ut, cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 1);
     CUN_UNITTEST_EVAL(ut, cb1.size() == 0);
     CUN_UNITTEST_NL(ut);
 
@@ -51,20 +54,26 @@ int main()
     CUN_UNITTEST_EVAL(ut, cb1.front() == 2);
     CUN_UNITTEST_EVAL(ut, cb1.back() == 2);
     CUN_UNITTEST_EVAL(ut, !cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 0);
     CUN_UNITTEST_EVAL(ut, cb1.size() == 1);
     CUN_UNITTEST_NL(ut);
 
     CUN_UNITTEST_NAME(ut, "drop data");
     CUN_UNITTEST_EVAL(ut, cb1.drop());
     CUN_UNITTEST_EVAL(ut, cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 1);
     CUN_UNITTEST_EVAL(ut, cb1.size() == 0);
     CUN_UNITTEST_NL(ut);
 
     CUN_UNITTEST_NAME(ut, "clear all data");
     CUN_UNITTEST_EVAL(ut, cb1.push(4));
     CUN_UNITTEST_EVAL(ut, !cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 0);
+    CUN_UNITTEST_EVAL(ut, cb1.size() == 1);
     CUN_UNITTEST_EXEC(ut, cb1.clear());
     CUN_UNITTEST_EVAL(ut, cb1.empty());
+    CUN_UNITTEST_EVAL(ut, cb1.free_size() == 1);
+    CUN_UNITTEST_EVAL(ut, cb1.size() == 0);
     CUN_UNITTEST_NL(ut);
 
     return EXIT_SUCCESS;