+++ ChangeLog.md
@@ -7,6 +7,7 @@ ChangeLog
 ### Added
 
 * Library: Circular buffer: drop: The first implementation.
+* Library: Circular buffer: free_size: The first implementation.
 
 [0.0.0.2025101901] - 2025-10-19
 -------------------------------